+++ src/main/java/com/jsk/schedule/domain/notification/entity/NotificationActionType.java
@@ -1,0 +1,7 @@
+package com.jsk.schedule.domain.notification.entity;
+
+public enum NotificationActionType {
+    CREATED,
+    UPDATED,
+    DELETED
+}